+++ Changelog.md
@@ -7,7 +7,13 @@ This project does _not_ adhere to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but contrary to it
 every new version is a new major version.
 
-## Unreleased
+## 9 - 2023-02-16
+
+### Added
+
+-   Masonry Layout
+
+## 8 - 2023-02-15
 
 ### Changed